+++ src/db/schemas/Game.js
@@ -40,6 +40,14 @@ const gameOptions = {
                 max: 5,
             },
         },
+        raven: {
+            markPenalty: {
+                type: Number,
+                default: 2,
+                min: 1,
+                max: 5,
+            },
+        },
     },
 };