+++ tests/e2e/game/game-options.test.js
@@ -17,6 +17,7 @@ const originalPlayers = [
     { name: "Dug", role: "three-brothers" },
     { name: "Dyg", role: "three-brothers" },
     { name: "Dπg", role: "villager" },
+    { name: "Dœg", role: "raven" },
 ];
 let token, game, players;
 
@@ -44,7 +45,7 @@ describe("K - Game options", () => {
                 done();
             });
     });
-    it("🎲 Creates game with brothers and sisters waking up every night, sheriff has regular vote and seer is talkative with JWT auth (POST /games)", done => {
+    it("🎲 Creates game with brothers and sisters waking up every night, sheriff has regular vote, seer is talkative and raven penalty to 3 with JWT auth (POST /games)", done => {
         chai.request(app)
             .post("/games")
             .set({ Authorization: `Bearer ${token}` })
@@ -56,6 +57,7 @@ describe("K - Game options", () => {
                         seer: { isTalkative: false },
                         twoSisters: { wakingUpInterval: 1 },
                         threeBrothers: { wakingUpInterval: 1 },
+                        raven: { markPenalty: 3 },
                     },
                 },
             })
@@ -66,6 +68,7 @@ describe("K - Game options", () => {
                 expect(game.options.roles.seer.isTalkative).to.be.false;
                 expect(game.options.roles.twoSisters.wakingUpInterval).to.equals(1);
                 expect(game.options.roles.threeBrothers.wakingUpInterval).to.equals(1);
+                expect(game.options.roles.raven.markPenalty).to.equals(3);
                 done();
             });
     });
@@ -108,6 +111,21 @@ describe("K - Game options", () => {
                 done();
             });
     });
+    it("🪶 Raven marks the werewolf (POST /games/:id/play)", done => {
+        players = game.players;
+        chai.request(app)
+            .post(`/games/${game._id}/play`)
+            .set({ Authorization: `Bearer ${token}` })
+            .send({ source: "raven", action: "mark", targets: [{ player: players[0]._id }] })
+            .end((err, res) => {
+                expect(res).to.have.status(200);
+                game = res.body;
+                expect(game.players[0].attributes).to.deep.include({ name: "raven-marked", source: "raven", remainingPhases: 2 });
+                expect(game.history[0].play.targets).to.exist;
+                expect(game.history[0].play.targets[0].player._id).to.equals(players[0]._id);
+                done();
+            });
+    });
     it("🐺 Werewolf eats the villager (POST /games/:id/play)", done => {
         players = game.players;
         chai.request(app)
@@ -124,7 +142,7 @@ describe("K - Game options", () => {
         expect(game.phase).to.equals("day");
         done();
     });
-    it("👪 Werewolf (sheriff) votes for one brother and the brother votes for the werewolf (POST /games/:id/play)", done => {
+    it("👪 Werewolf (sheriff) and three other players votes for one brother and the brother votes for the werewolf (POST /games/:id/play)", done => {
         players = game.players;
         chai.request(app)
             .post(`/games/${game._id}/play`)
@@ -132,6 +150,9 @@ describe("K - Game options", () => {
             .send({
                 source: "all", action: "vote", votes: [
                     { from: players[0]._id, for: players[5]._id },
+                    { from: players[1]._id, for: players[5]._id },
+                    { from: players[2]._id, for: players[5]._id },
+                    { from: players[3]._id, for: players[5]._id },
                     { from: players[5]._id, for: players[0]._id },
                 ],
             })
@@ -194,6 +215,18 @@ describe("K - Game options", () => {
                 done();
             });
     });
+    it("🪶 Raven skips (POST /games/:id/play)", done => {
+        players = game.players;
+        chai.request(app)
+            .post(`/games/${game._id}/play`)
+            .set({ Authorization: `Bearer ${token}` })
+            .send({ source: "raven", action: "mark" })
+            .end((err, res) => {
+                expect(res).to.have.status(200);
+                game = res.body;
+                done();
+            });
+    });
     it("🐺 Werewolf eats one of the two sisters (POST /games/:id/play)", done => {
         players = game.players;
         chai.request(app)
@@ -228,8 +261,8 @@ describe("K - Game options", () => {
         expect(game.turn).to.equals(3);
         done();
     });
-    it("🎲 Game is waiting for 'werewolves' to 'eat' because sisters and brothers are all alone", done => {
-        expect(game.waiting[0]).to.deep.equals({ for: "werewolves", to: "eat" });
+    it("🎲 Game is waiting for 'raven' to 'mark' because sisters and brothers are all alone", done => {
+        expect(game.waiting[0]).to.deep.equals({ for: "raven", to: "mark" });
         done();
     });
     it("🎲 Cancels game (PATCH /games/:id)", done => {
@@ -300,6 +333,18 @@ describe("K - Game options", () => {
                 done();
             });
     });
+    it("🪶 Raven skips (POST /games/:id/play)", done => {
+        players = game.players;
+        chai.request(app)
+            .post(`/games/${game._id}/play`)
+            .set({ Authorization: `Bearer ${token}` })
+            .send({ source: "raven", action: "mark" })
+            .end((err, res) => {
+                expect(res).to.have.status(200);
+                game = res.body;
+                done();
+            });
+    });
     it("🐺 Werewolf eats the villager (POST /games/:id/play)", done => {
         players = game.players;
         chai.request(app)
@@ -334,8 +379,8 @@ describe("K - Game options", () => {
         expect(game.turn).to.equals(2);
         done();
     });
-    it("🎲 Game is waiting for 'werewolves' to 'eat' because brothers and sisters are never waking up again according to game options", done => {
-        expect(game.waiting[0]).to.deep.equals({ for: "werewolves", to: "eat" });
+    it("🎲 Game is waiting for 'raven' to 'mark' because brothers and sisters are never waking up again according to game options", done => {
+        expect(game.waiting[0]).to.deep.equals({ for: "raven", to: "mark" });
         done();
     });
     it("🎲 Cancels game (PATCH /games/:id)", done => {
@@ -350,7 +395,7 @@ describe("K - Game options", () => {
                 done();
             });
     });
-    it("🎲 Creates game with disabled sheriff option and brothers and sisters waking up only the first night with JWT auth (POST /games)", done => {
+    it("🎲 Creates game with disabled sheriff option, brothers and sisters waking up only the first night and raven penalty to 1 with JWT auth (POST /games)", done => {
         chai.request(app)
             .post("/games")
             .set({ Authorization: `Bearer ${token}` })
@@ -361,6 +406,7 @@ describe("K - Game options", () => {
                         sheriff: { enabled: false },
                         twoSisters: { wakingUpInterval: 0 },
                         threeBrothers: { wakingUpInterval: 0 },
+                        raven: { markPenalty: 1 },
                     },
                 },
             })
@@ -390,6 +436,18 @@ describe("K - Game options", () => {
             .end((err, res) => {
                 game = res.body;
                 expect(res).to.have.status(200);
+                done();
+            });
+    });
+    it("🪶 Raven mark a brother (POST /games/:id/play)", done => {
+        players = game.players;
+        chai.request(app)
+            .post(`/games/${game._id}/play`)
+            .set({ Authorization: `Bearer ${token}` })
+            .send({ source: "raven", action: "mark", targets: [{ player: players[5]._id }] })
+            .end((err, res) => {
+                expect(res).to.have.status(200);
+                game = res.body;
                 done();
             });
     });
@@ -425,6 +483,7 @@ describe("K - Game options", () => {
                 game = res.body;
                 expect(game.players[0].isAlive).to.be.true;
                 expect(game.players[1].isAlive).to.be.true;
+                expect(game.players[5].isAlive).to.be.true;
                 done();
             });
     });
@@ -448,7 +507,7 @@ describe("K - Game options", () => {
                 done();
             });
     });
-    it("👪 Another tie in votes between the werewolf and one of the two sisters, then nobody dies (POST /games/:id/play)", done => {
+    it("👪 Another tie in votes between the werewolf, one of the two sisters and one of the three brothers, then nobody dies (POST /games/:id/play)", done => {
         players = game.players;
         chai.request(app)
             .post(`/games/${game._id}/play`)
@@ -459,6 +518,8 @@ describe("K - Game options", () => {
                     { from: players[1]._id, for: players[0]._id },
                     { from: players[2]._id, for: players[1]._id },
                     { from: players[3]._id, for: players[0]._id },
+                    { from: players[4]._id, for: players[5]._id },
+                    { from: players[7]._id, for: players[5]._id },
                 ],
             })
             .end((err, res) => {
@@ -466,6 +527,7 @@ describe("K - Game options", () => {
                 game = res.body;
                 expect(game.players[0].isAlive).to.be.true;
                 expect(game.players[1].isAlive).to.be.true;
+                expect(game.players[5].isAlive).to.be.true;
                 expect(game.history[0].deadPlayers).to.not.exist;
                 done();
             });
@@ -474,6 +536,18 @@ describe("K - Game options", () => {
         expect(game.phase).to.equals("night");
         expect(game.turn).to.equals(2);
         done();
+    });
+    it("🪶 Raven skips (POST /games/:id/play)", done => {
+        players = game.players;
+        chai.request(app)
+            .post(`/games/${game._id}/play`)
+            .set({ Authorization: `Bearer ${token}` })
+            .send({ source: "raven", action: "mark" })
+            .end((err, res) => {
+                expect(res).to.have.status(200);
+                game = res.body;
+                done();
+            });
     });
     it("🐺 Werewolf eats one of the three brothers (POST /games/:id/play)", done => {
         players = game.players;